trying to break the logic of NotificationProvider & toast id bugfix
