--- src/utilities/eventEmitter.js
@@ -1,4 +1,0 @@
-import mitt from 'mitt';
-
-const eventEmitter = mitt();
-export default eventEmitter;
