+++ src/utilities/notifications/eventEmitter.js
@@ -1,0 +1,3 @@
+import mitt from 'mitt';
+
+export const eventEmitter = mitt();
