+++ src/utilities/notifications/notificationContext.js
@@ -1,0 +1,4 @@
+import { createContext, useContext } from 'react';
+
+export const NotificationContext = createContext({});
+export const useNotifications = () => useContext(NotificationContext);
