+++ src/utilities/notifications/notifyExternally.js
@@ -1,0 +1,5 @@
+import {eventEmitter} from './eventEmitter';
+
+export const notifyExternally = (headMessage, bodyMessage, duration = 5000) => {
+    eventEmitter.emit('notify', { headMessage, bodyMessage, duration });
+};
